separated out Models and added seed function
