+++ server/db/Models/Movie.js
@@ -1,0 +1,9 @@
+const Sequelize = require('sequelize');
+const chalk = require('chalk');
+const { UUID, UUIDV4, STRING, INTEGER } = Sequelize;
+
+const Movie = db.define('Vendor', {
+    // Nick to work on this //
+})
+
+export default Movie;
